test: 테스트 코드 API 멀티모듈로 이동
