+++ Api/src/test/java/com/mapz/api/common/fixtures/DiaryImageFixtures.java
@@ -1,0 +1,14 @@
+package com.mapz.api.common.fixtures;
+
+import com.mapz.domain.domains.diary.entity.Diary;
+import com.mapz.domain.domains.diary.entity.DiaryImage;
+
+public class DiaryImageFixtures {
+
+    public static DiaryImage createDiaryImage(Diary diary) {
+        return DiaryImage.builder()
+                .diary(diary)
+                .diaryImageUrl("url")
+                .build();
+    }
+}
