+++ Api/src/test/java/com/mapz/api/common/fixtures/LikeFixtures.java
@@ -1,0 +1,5 @@
+package com.mapz.api.common.fixtures;
+
+public class LikeFixtures {
+
+}
