--- src/test/java/com/cheocharm/MapZ/common/fixtures/LikeFixtures.java
@@ -1,5 +1,0 @@
-package com.cheocharm.MapZ.common.fixtures;
-
-public class LikeFixtures {
-
-}
